requirements.txt
Flask==3.0.2
Flask-SQLAlchemy==3.1.1
Flask-MySQLdb==2.0.0
mysql-connector-python==8.3.0
python-dotenv==1.0.1
Flask-Login==0.6.3
Werkzeug==3.0.1
PyJWT==2.8.0
python-dateutil==2.8.2 